Move modelform customisation logic to wagtail.admin.forms.models
